refactor(callbacks): extract Overview briefing surface to _callbacks_overview.py
Step 7c of the callbacks.py decomposition (issue #87). Third and final
Overview sub-PR. Stacked on PR #105.

## What moves

12 functions from the Overview briefing surface — covering the
sparkline, the AI briefing card, weather context cards, the "since
your last visit" diff card, data-health badges, persona-aware
spotlight charts, the cross-tab insight digest, the news feed, and
the persona-aware KPI row — all move out of callbacks.py and join
the rest of the Overview helpers in
``components/_callbacks_overview.py``:

* ``_build_overview_sparkline`` — 24h demand sparkline (also the
  default spotlight chart for the grid_ops persona).
* ``_build_overview_briefing`` — AI executive briefing card with
  observations list and "AI Analysis" / "Data Summary" source label.
* ``_build_weather_context`` — 4-up weather KPI cards (temperature,
  wind, humidity, cloud cover) with semantic color thresholds.
* ``_build_changes_card`` — "Since your last visit" diff card driven
  by the session-snapshot store.
* ``_build_overview_data_health`` — per-source freshness badges
  (EIA Demand, Weather, NOAA Alerts) using the ``data-freshness-store``.
* ``_build_overview_spotlight`` + ``_spotlight_renewables`` +
  ``_spotlight_trader`` + ``_spotlight_model_accuracy`` — persona-
  aware spotlight dispatcher and three persona-specific chart builders.
* ``_build_overview_digest`` — cross-tab insight aggregator pulling
  from ``_PREDICTION_CACHE`` / ``_BACKTEST_CACHE`` / ``_GENERATION_CACHE``.
* ``_build_overview_news`` — energy news feed with demo-data fallback.
* ``_build_persona_kpis`` — 4-up persona-tailored KPI row (Grid Ops /
  Renewables / Trader / Data Scientist), with Redis fallback for
  every metric source.

## Public-import surface preserved

callbacks.py re-imports all 12 functions by name, with explicit
``# noqa: F401`` markers on the 10 that are re-exports for tests
rather than internally called. The two callbacks.py STILL calls
directly — ``_build_weather_context`` (line 2301) and
``_build_persona_kpis`` (via tests) — keep their imports without
markers.

Also re-added ``_GENERATION_CACHE`` to the shared re-export block
with a ``# noqa`` marker — the test fixture
``_clear_module_caches`` calls ``cb._GENERATION_CACHE.clear()``
through the callbacks.py namespace, but callbacks.py itself no
longer references it directly.

## Test patches updated

``redis_get`` and ``_BACKTEST_CACHE`` patches inside the two
``TestBuildPersonaKpis`` classes (one each in
``test_callbacks_helpers.py`` and ``test_callbacks_module_helpers.py``)
rewired from ``components.callbacks.X`` to
``components._callbacks_overview.X`` — ``_build_persona_kpis`` now
reads those names from its own module's namespace.

Patches in earlier classes (``TestRunBacktestForHorizon``,
``TestRunForecastOutlook``, etc.) intentionally left as
``components.callbacks.X`` since those functions still live in
callbacks.py.

## Size impact

callbacks.py: 4,686 → 3,750 (–936 lines)
_callbacks_overview.py: 1,089 → 2,057 (+968 lines)

**callbacks.py crosses below 4,000 lines for the first time.**

## Cumulative progress

* PR #98 (shared):       7,167 → 7,051 (–116)
* PR #99 (us_grid):      7,051 → 6,389 (–662)
* PR #100 (models):      6,389 → 6,183 (–206)
* PR #101 (alerts):      6,183 → 5,978 (–205)
* PR #102 (generation):  5,978 → 5,802 (–176)
* PR #103 (weather):     5,802 → 5,668 (–134)
* PR #104 (overview 7a): 5,668 → 5,369 (–299)
* PR #105 (overview 7b): 5,369 → 4,686 (–683)
* this PR (overview 7c): 4,686 → 3,750 (–936)

Total: −3,417 lines from the monolith across 9 PRs (−47.7%).

## Tests

All 1,568 unit tests pass locally. Ruff clean.
