+++ components/callbacks.py
@@ -48,23 +48,35 @@ from components._callbacks_models import (
     _models_tab_from_redis,
 )
 from components._callbacks_overview import (
+    _build_changes_card,  # noqa: F401 — re-export (tests + helper-callable surface)
     _build_drivers_panel,
     _build_generation_panel,
     _build_models_leaderboard,
+    _build_overview_briefing,  # noqa: F401 — re-export (tests + helper-callable surface)
+    _build_overview_data_health,  # noqa: F401 — re-export (tests/unit/test_tab_overview.py)
+    _build_overview_digest,  # noqa: F401 — re-export (tests/unit/test_tab_overview.py)
     _build_overview_hero_chart,
     _build_overview_insight,
     _build_overview_metrics_items,
     _build_overview_model_card,
+    _build_overview_news,  # noqa: F401 — re-export (tests/unit/test_tab_overview.py)
+    _build_overview_sparkline,  # noqa: F401 — re-export (tests/unit/test_tab_overview.py)
+    _build_overview_spotlight,  # noqa: F401 — re-export (tests/unit/test_tab_overview.py)
     _build_overview_title,
+    _build_persona_kpis,  # noqa: F401 — re-export (tests/unit/test_callbacks_helpers.py)
     _build_risk_insight,
     _build_scenarios_panel,
+    _build_weather_context,
     _fetch_generation_cached,  # noqa: F401 — re-export (tests/unit/test_callbacks_*)
+    _spotlight_model_accuracy,  # noqa: F401 — re-export (tests/unit/test_tab_overview.py)
+    _spotlight_renewables,  # noqa: F401 — re-export (tests/unit/test_tab_overview.py)
+    _spotlight_trader,  # noqa: F401 — re-export (tests/unit/test_tab_overview.py)
 )
 from components._callbacks_shared import (
     _BACKTEST_CACHE,
     _CACHE_VERSION,
     _EIA_FUEL_MAP,  # noqa: F401 — re-export (tests/unit/test_callbacks_helpers.py::TestModuleConstants)
-    _GENERATION_CACHE,
+    _GENERATION_CACHE,  # noqa: F401 — re-export (test fixture `_clear_module_caches`)
     _MAP_BORDER_COLOR,  # noqa: F401 — re-export (tests/unit/test_us_grid_*)
     _MAP_COLORSCALE,  # noqa: F401 — re-export (tests/unit/test_us_grid_*)
     _MAP_LAND_COLOR,  # noqa: F401 — re-export (tests/unit/test_us_grid_*)
@@ -85,19 +97,16 @@ from components._callbacks_shared import (
 from components._callbacks_weather import (
     _weather_tab_from_redis,  # noqa: F401 — re-export (tests/unit/test_redis_fast_paths.py); fast path is currently orphaned in register_callbacks
 )
-from components.accessibility import CB_PALETTE, LINE_STYLES
+from components.accessibility import LINE_STYLES
 from components.cards import (
     build_alert_card,
-    build_kpi_row,
     build_metrics_bar,
     build_model_metrics_card,
-    build_news_feed,
     build_page_title,
 )
 from config import (
     CACHE_TTL_SECONDS,
     EIA_API_KEY,
-    REGION_CAPACITY_MW,
     REGION_NAMES,
     REQUIRE_REDIS,
     WEATHER_VARIABLES,
@@ -3323,951 +3332,6 @@ def register_callbacks(app):
 # Companion to the ``> 0`` filter in ``_build_overview_metrics_items`` —
 # this formalizes that pattern as the single source of truth, so future
 # tabs (US Grid, Risk, Models) can't accidentally re-introduce the bug.
-
-
-def _build_overview_sparkline(demand_df: pd.DataFrame | None, region: str) -> go.Figure:
-    """Build a compact 24h demand sparkline for the overview tab."""
-    if demand_df is None or demand_df.empty or "demand_mw" not in demand_df.columns:
-        return _empty_figure("No demand data")
-
-    df = demand_df.copy()
-    df["timestamp"] = pd.to_datetime(df["timestamp"])
-    last_24h = df.tail(24)
-
-    if last_24h.empty:
-        return _empty_figure("No recent demand data")
-
-    fig = go.Figure()
-    fig.add_trace(
-        go.Scatter(
-            x=last_24h["timestamp"],
-            y=last_24h["demand_mw"],
-            mode="lines",
-            line=dict(color=CB_PALETTE["blue"], width=2),
-            fill="tozeroy",
-            fillcolor="rgba(0,114,178,0.15)",
-            name="Demand",
-            hovertemplate="%{x|%H:%M}<br>%{y:,.0f} MW<extra></extra>",
-        )
-    )
-    sparkline_layout = _layout(
-        uirevision=region,
-        showlegend=False,
-        margin=dict(l=40, r=10, t=10, b=30),
-    )
-    fig.update_layout(
-        **sparkline_layout,
-        xaxis=dict(
-            showgrid=False,
-            tickformat="%H:%M",
-        ),
-        yaxis=dict(
-            showgrid=True,
-            gridcolor="rgba(255,255,255,0.05)",
-            tickformat=",.0f",
-            title="MW",
-        ),
-    )
-    return fig
-
-
-def _build_overview_briefing(
-    persona_id: str,
-    region: str,
-    demand_df: pd.DataFrame | None,
-    weather_df: pd.DataFrame | None,
-) -> html.Div:
-    """Build the AI executive briefing section."""
-    from data.ai_briefing import generate_briefing
-
-    try:
-        result = generate_briefing(persona_id, region, demand_df, weather_df)
-    except Exception as exc:
-        log.error("overview_briefing_failed", error=str(exc))
-        return html.Div(
-            "Briefing unavailable",
-            style={"color": "#A8B3C7", "fontStyle": "italic"},
-        )
-
-    persona = get_persona(persona_id)
-
-    children = [
-        html.P(
-            result.summary,
-            style={
-                "color": "#DDE6F2",
-                "fontSize": "0.9rem",
-                "lineHeight": "1.6",
-                "marginBottom": "12px",
-            },
-        ),
-    ]
-
-    if result.observations:
-        obs_items = []
-        for obs in result.observations:
-            obs_items.append(
-                html.Li(
-                    obs,
-                    style={
-                        "color": "#A8B3C7",
-                        "fontSize": "0.82rem",
-                        "marginBottom": "4px",
-                        "lineHeight": "1.5",
-                    },
-                )
-            )
-        children.append(
-            html.Ul(
-                obs_items,
-                style={"paddingLeft": "20px", "marginBottom": "8px"},
-            )
-        )
-
-    source_label = "AI Analysis" if result.source == "claude" else "Data Summary"
-    children.append(
-        html.Span(
-            source_label,
-            style={
-                "fontSize": "0.65rem",
-                "color": "#A8B3C7",
-                "textTransform": "uppercase",
-                "letterSpacing": "0.5px",
-            },
-        )
-    )
-
-    return html.Div(
-        children,
-        style={"borderLeft": f"4px solid {persona.color}"},
-        className="briefing-card-content",
-    )
-
-
-def _build_weather_context(latest: "pd.Series") -> html.Div:
-    """Build a row of weather KPI mini-cards from the latest weather reading."""
-    import dash_bootstrap_components as dbc
-
-    temp = latest.get("temperature_2m")
-    wind = latest.get("wind_speed_80m", latest.get("wind_speed_10m"))
-    humidity = latest.get("relative_humidity_2m")
-    cloud = latest.get("cloud_cover")
-
-    cards = []
-
-    if temp is not None:
-        t = float(temp)
-        color = "#FF5C7A" if t >= 95 else ("#FFB84D" if t >= 85 else "#2BD67B")
-        cards.append(
-            dbc.Col(
-                html.Div(
-                    [
-                        html.P("TEMPERATURE", className="kpi-label"),
-                        html.H4(
-                            f"{t:.0f}°F",
-                            className="kpi-value",
-                            style={"fontSize": "1.3rem"},
-                        ),
-                    ],
-                    className="kpi-card",
-                    style={"borderTop": f"3px solid {color}"},
-                ),
-                md=3,
-            )
-        )
-
-    if wind is not None:
-        w = float(wind)
-        color = "#FF5C7A" if w >= 40 else ("#FFB84D" if w >= 25 else "#2BD67B")
-        cards.append(
-            dbc.Col(
-                html.Div(
-                    [
-                        html.P("WIND SPEED", className="kpi-label"),
-                        html.H4(
-                            f"{w:.0f} mph",
-                            className="kpi-value",
-                            style={"fontSize": "1.3rem"},
-                        ),
-                    ],
-                    className="kpi-card",
-                    style={"borderTop": f"3px solid {color}"},
-                ),
-                md=3,
-            )
-        )
-
-    if humidity is not None:
-        h = float(humidity)
-        color = "#FFB84D" if h >= 80 else "#2BD67B"
-        cards.append(
-            dbc.Col(
-                html.Div(
-                    [
-                        html.P("HUMIDITY", className="kpi-label"),
-                        html.H4(
-                            f"{h:.0f}%",
-                            className="kpi-value",
-                            style={"fontSize": "1.3rem"},
-                        ),
-                    ],
-                    className="kpi-card",
-                    style={"borderTop": f"3px solid {color}"},
-                ),
-                md=3,
-            )
-        )
-
-    if cloud is not None:
-        c = float(cloud)
-        color = "#A8B3C7"
-        cards.append(
-            dbc.Col(
-                html.Div(
-                    [
-                        html.P("CLOUD COVER", className="kpi-label"),
-                        html.H4(
-                            f"{c:.0f}%",
-                            className="kpi-value",
-                            style={"fontSize": "1.3rem"},
-                        ),
-                    ],
-                    className="kpi-card",
-                    style={"borderTop": f"3px solid {color}"},
-                ),
-                md=3,
-            )
-        )
-
-    if not cards:
-        return html.Div()
-
-    return dbc.Row(cards, className="g-2")
-
-
-def _build_changes_card(
-    changes_json: str | None,
-    persona: str,
-    region: str,
-    snapshots: dict | None,
-) -> html.Div:
-    """Build the 'Since your last visit' card for the Overview tab (NEXD-8)."""
-    import json as _json
-
-    from data.session_diff import format_relative_time
-
-    if not changes_json:
-        return html.Div()
-
-    try:
-        changes = _json.loads(changes_json) if isinstance(changes_json, str) else changes_json
-    except Exception:
-        return html.Div()
-
-    if not changes:
-        return html.Div()
-
-    # Get previous visit timestamp for this region
-    prev_timestamp = None
-    if snapshots and isinstance(snapshots, dict):
-        entry = snapshots.get(region)
-        if entry and isinstance(entry, dict):
-            prev_timestamp = entry.get("timestamp")
-
-    rel_time = format_relative_time(prev_timestamp) if prev_timestamp else ""
-
-    # Build bullet items
-    items = []
-    for change in changes[:5]:
-        icon = change.get("icon", "")
-        text = change.get("text", "")
-        items.append(
-            html.Div(
-                [
-                    html.Span(icon, className="change-icon"),
-                    html.Span(text),
-                ],
-                className="change-item",
-            )
-        )
-
-    header_parts = [
-        html.Span("Since your last visit", className="changes-title"),
-    ]
-    if rel_time:
-        header_parts.append(html.Span(rel_time, className="changes-timestamp"))
-
-    return html.Div(
-        [
-            html.Div(header_parts, className="changes-header"),
-            *items,
-        ],
-        className="changes-card",
-    )
-
-
-def _build_overview_data_health(freshness_data: dict | None) -> html.Div:
-    """Build data health badges showing per-source freshness."""
-    if not freshness_data:
-        return html.Div()
-
-    source_config = {
-        "demand": {"label": "EIA Demand", "icon": "\u26a1"},
-        "weather": {"label": "Weather", "icon": "\u2601"},
-        "alerts": {"label": "NOAA Alerts", "icon": "\u26a0"},
-    }
-
-    status_colors = {
-        "fresh": "#2BD67B",
-        "stale": "#FFB84D",
-        "demo": "#A8B3C7",
-        "error": "#FF5C7A",
-    }
-
-    badges = []
-    for source, status in freshness_data.items():
-        if source == "timestamp":
-            continue
-        cfg = source_config.get(source, {"label": source.title(), "icon": "\u25cf"})
-        color = status_colors.get(status, "#A8B3C7")
-        status_text = status.upper() if status != "fresh" else "LIVE"
-        badges.append(
-            html.Div(
-                [
-                    html.Span(cfg["icon"], style={"marginRight": "6px"}),
-                    html.Span(
-                        cfg["label"],
-                        style={"fontWeight": "600", "marginRight": "6px"},
-                    ),
-                    html.Span(
-                        status_text,
-                        style={
-                            "fontSize": "0.65rem",
-                            "padding": "1px 6px",
-                            "borderRadius": "3px",
-                            "background": f"{color}20",
-                            "color": color,
-                        },
-                    ),
-                ],
-                className="data-health-badge",
-                style={
-                    "display": "inline-flex",
-                    "alignItems": "center",
-                    "fontSize": "0.75rem",
-                    "color": "#A8B3C7",
-                    "padding": "4px 12px",
-                    "marginRight": "12px",
-                },
-            )
-        )
-
-    if not badges:
-        return html.Div()
-
-    return html.Div(
-        [
-            html.Span(
-                "DATA SOURCES",
-                style={
-                    "fontSize": "0.65rem",
-                    "color": "#A8B3C7",
-                    "textTransform": "uppercase",
-                    "letterSpacing": "1px",
-                    "marginRight": "16px",
-                },
-            ),
-            *badges,
-        ],
-        style={
-            "display": "flex",
-            "flexWrap": "wrap",
-            "alignItems": "center",
-            "padding": "8px 12px",
-            "background": "#11182D",
-            "borderRadius": "6px",
-        },
-    )
-
-
-def _build_overview_spotlight(
-    persona_id: str,
-    region: str,
-    demand_df: pd.DataFrame | None,
-    weather_df: pd.DataFrame | None,
-) -> go.Figure:
-    """Build persona-specific spotlight chart for overview."""
-    if persona_id == "renewables":
-        return _spotlight_renewables(weather_df, region)
-    if persona_id == "trader":
-        return _spotlight_trader(demand_df, region)
-    if persona_id == "data_scientist":
-        return _spotlight_model_accuracy(region)
-    # Default: grid_ops → demand sparkline
-    return _build_overview_sparkline(demand_df, region)
-
-
-def _spotlight_renewables(weather_df: pd.DataFrame | None, region: str) -> go.Figure:
-    """Renewable generation potential chart."""
-    if weather_df is None or weather_df.empty:
-        return _empty_figure("No weather data for renewable outlook")
-
-    df = weather_df.copy()
-    if "timestamp" in df.columns:
-        df["timestamp"] = pd.to_datetime(df["timestamp"])
-    last_48h = df.tail(48)
-
-    fig = go.Figure()
-    if "wind_speed_80m" in last_48h.columns:
-        fig.add_trace(
-            go.Scatter(
-                x=last_48h.get("timestamp", list(range(len(last_48h)))),
-                y=last_48h["wind_speed_80m"],
-                mode="lines",
-                line=dict(color=CB_PALETTE.get("sky", "#56B4E9"), width=2),
-                name="Wind (mph)",
-                hovertemplate="%{y:.0f} mph<extra>Wind</extra>",
-            )
-        )
-    if "shortwave_radiation" in last_48h.columns:
-        fig.add_trace(
-            go.Scatter(
-                x=last_48h.get("timestamp", list(range(len(last_48h)))),
-                y=last_48h["shortwave_radiation"],
-                mode="lines",
-                line=dict(color=CB_PALETTE.get("orange", "#E69F00"), width=2),
-                name="Solar (W/m\u00b2)",
-                yaxis="y2",
-                hovertemplate="%{y:.0f} W/m\u00b2<extra>Solar</extra>",
-            )
-        )
-
-    renew_layout = _layout(
-        uirevision=region,
-        margin=dict(l=45, r=45, t=35, b=40),
-        legend=dict(orientation="h", y=-0.15, font=dict(size=10)),
-    )
-    fig.update_layout(
-        **renew_layout,
-        title=dict(text="Renewable Potential (48h)", font=dict(size=13, color="#DDE6F2")),
-        showlegend=True,
-        yaxis=dict(
-            title="Wind (mph)",
-            showgrid=True,
-            gridcolor="rgba(255,255,255,0.05)",
-        ),
-        yaxis2=dict(
-            title="Solar (W/m\u00b2)",
-            overlaying="y",
-            side="right",
-            showgrid=False,
-        ),
-        xaxis=dict(showgrid=False, tickformat="%b %d %H:%M"),
-    )
-    return fig
-
-
-def _spotlight_trader(demand_df: pd.DataFrame | None, region: str) -> go.Figure:
-    """Demand vs capacity utilization chart for traders."""
-    capacity = REGION_CAPACITY_MW.get(region, 50000)
-
-    if demand_df is None or demand_df.empty:
-        return _empty_figure("No demand data for market view")
-
-    df = demand_df.copy()
-    if "timestamp" in df.columns:
-        df["timestamp"] = pd.to_datetime(df["timestamp"])
-    last_48h = df.tail(48)
-
-    fig = go.Figure()
-    fig.add_trace(
-        go.Scatter(
-            x=last_48h.get("timestamp", list(range(len(last_48h)))),
-            y=last_48h["demand_mw"],
-            mode="lines",
-            line=dict(color=CB_PALETTE["blue"], width=2),
-            fill="tozeroy",
-            fillcolor="rgba(0,114,178,0.15)",
-            name="Demand",
-            hovertemplate="%{y:,.0f} MW<extra>Demand</extra>",
-        )
-    )
-
-    # Capacity line
-    fig.add_hline(
-        y=capacity,
-        line_dash="dot",
-        line_color="#FF5C7A",
-        annotation_text=f"Capacity: {capacity:,.0f} MW",
-        annotation_position="top left",
-        annotation_font_size=10,
-        annotation_font_color="#FF5C7A",
-    )
-
-    # Pricing tier thresholds
-    for pct, label, color in [
-        (0.85, "High tier (85%)", "#FFB84D"),
-        (0.70, "Moderate (70%)", "#A8B3C7"),
-    ]:
-        fig.add_hline(
-            y=capacity * pct,
-            line_dash="dot",
-            line_color=color,
-            line_width=1,
-            annotation_text=label,
-            annotation_position="bottom left",
-            annotation_font_size=9,
-            annotation_font_color=color,
-        )
-
-    trader_layout = _layout(uirevision=region, margin=dict(l=50, r=10, t=35, b=30))
-    fig.update_layout(
-        **trader_layout,
-        title=dict(text="Demand vs Capacity", font=dict(size=13, color="#DDE6F2")),
-        showlegend=False,
-        xaxis=dict(showgrid=False, tickformat="%b %d %H:%M"),
-        yaxis=dict(
-            showgrid=True,
-            gridcolor="rgba(255,255,255,0.05)",
-            tickformat=",.0f",
-            title="MW",
-        ),
-    )
-    return fig
-
-
-def _spotlight_model_accuracy(region: str) -> go.Figure:
-    """Model accuracy bar chart for data scientists."""
-    # Pull from backtest cache if available
-    models = ["prophet", "arima", "xgboost"]
-    mape_values = []
-
-    for model_name in models:
-        mape = None
-        for horizon in [168, 24, 720]:
-            bt_key = (region, horizon, model_name, DEFAULT_BACKTEST_EXOG_MODE)
-            if bt_key in _BACKTEST_CACHE:
-                result_dict, _, _ = _BACKTEST_CACHE[bt_key]
-                if isinstance(result_dict, dict) and "mape" in result_dict:
-                    mape = result_dict["mape"]
-                    break
-        mape_values.append(mape if mape is not None else 4.5 + len(model_name) * 0.3)
-
-    colors = [
-        CB_PALETTE.get("vermillion", "#D55E00"),
-        CB_PALETTE.get("blue", "#0072B2"),
-        CB_PALETTE.get("green", "#009E73"),
-    ]
-
-    fig = go.Figure()
-    fig.add_trace(
-        go.Bar(
-            x=[m.title() for m in models],
-            y=mape_values,
-            marker_color=colors,
-            text=[f"{v:.1f}%" for v in mape_values],
-            textposition="outside",
-            textfont=dict(color="#DDE6F2", size=11),
-            hovertemplate="%{x}: %{y:.2f}% MAPE<extra></extra>",
-        )
-    )
-
-    model_layout = _layout(uirevision=region, margin=dict(l=40, r=10, t=35, b=30))
-    fig.update_layout(
-        **model_layout,
-        title=dict(text="Model MAPE Comparison", font=dict(size=13, color="#DDE6F2")),
-        showlegend=False,
-        yaxis=dict(
-            title="MAPE (%)",
-            showgrid=True,
-            gridcolor="rgba(255,255,255,0.05)",
-        ),
-        xaxis=dict(showgrid=False),
-    )
-    return fig
-
-
-def _build_overview_digest(
-    persona_id: str,
-    region: str,
-    demand_df: pd.DataFrame | None,
-    weather_df: pd.DataFrame | None,
-) -> html.Div:
-    """Aggregate top insights from all tabs into a cross-tab digest."""
-    from components.insights import (
-        Insight,
-        build_insight_card,
-        generate_tab1_insights,
-        generate_tab2_insights,
-        generate_tab3_insights,
-        generate_tab4_insights,
-    )
-
-    all_insights: list[Insight] = []
-
-    # Tab 1: Historical demand insights
-    try:
-        tab1 = generate_tab1_insights(persona_id, region, demand_df, weather_df)
-        all_insights.extend(tab1)
-    except Exception:
-        pass
-
-    # Tab 2: Forecast insights (need predictions — skip if unavailable)
-    try:
-        for horizon in [168, 24]:
-            pred_key = (region, horizon)
-            if pred_key in _PREDICTION_CACHE:
-                preds, timestamps, _, _ = _PREDICTION_CACHE[pred_key]
-                tab2 = generate_tab2_insights(persona_id, region, preds, timestamps)
-                all_insights.extend(tab2)
-                break
-    except Exception:
-        pass
-
-    # Validation tab: model accuracy insights (from backtest cache)
-    try:
-        for model_name in ["xgboost", "prophet", "arima"]:
-            for horizon in [168, 24, 720]:
-                bt_key = (region, horizon, model_name, DEFAULT_BACKTEST_EXOG_MODE)
-                if bt_key in _BACKTEST_CACHE:
-                    result_dict, _, _ = _BACKTEST_CACHE[bt_key]
-                    if isinstance(result_dict, dict) and "mape" in result_dict:
-                        metrics = {model_name: result_dict}
-                        tab3 = generate_tab3_insights(
-                            persona_id,
-                            region,
-                            metrics,
-                            model_name=model_name,
-                            horizon_hours=horizon,
-                        )
-                        all_insights.extend(tab3)
-                        break
-            else:
-                continue
-            break
-    except Exception:
-        pass
-
-    # Tab 4: Grid insights (from generation cache)
-    try:
-        if region in _GENERATION_CACHE:
-            gen_df, _ = _GENERATION_CACHE[region]
-            if gen_df is not None and not gen_df.empty:
-                gen_copy = gen_df.copy()
-                gen_copy["timestamp"] = pd.to_datetime(gen_copy["timestamp"])
-                pivot = gen_copy.pivot_table(
-                    index="timestamp",
-                    columns="fuel_type",
-                    values="generation_mw",
-                    aggfunc="sum",
-                ).fillna(0)
-                total_gen = pivot.sum(axis=1)
-                renewable_cols = [
-                    c for c in pivot.columns if c.lower() in ("wind", "solar", "hydro")
-                ]
-                renewable_gen = (
-                    pivot[renewable_cols].sum(axis=1) if renewable_cols else total_gen * 0
-                )
-                renewable_pct = (
-                    (renewable_gen.sum() / total_gen.sum() * 100) if total_gen.sum() > 0 else 0.0
-                )
-                net_load = total_gen - renewable_gen
-                tab4 = generate_tab4_insights(
-                    persona_id=persona_id,
-                    region=region,
-                    net_load=net_load,
-                    demand=total_gen,
-                    renewable_pct=renewable_pct,
-                    pivot=pivot,
-                    timestamps=pd.DatetimeIndex(pivot.index),
-                )
-                all_insights.extend(tab4)
-    except Exception:
-        pass
-
-    if not all_insights:
-        return html.Div(
-            html.P(
-                "No insights available yet. Explore tabs to generate data.",
-                style={"color": "#A8B3C7", "fontSize": "0.82rem", "fontStyle": "italic"},
-            )
-        )
-
-    # Sort by severity (warning first)
-    severity_order = {"warning": 0, "notable": 1, "info": 2}
-    all_insights.sort(key=lambda i: severity_order.get(i.severity, 2))
-
-    return build_insight_card(all_insights, persona_id, "Overview", max_insights=5)
-
-
-def _build_overview_news() -> html.Div:
-    """Fetch and render energy news for the overview tab."""
-    from data.news_client import fetch_energy_news
-
-    try:
-        articles = fetch_energy_news(page_size=10)
-        if not articles:
-            from data.news_client import _get_demo_news
-
-            articles = _get_demo_news()
-        return build_news_feed(articles)
-    except Exception as e:
-        log.error("overview_news_failed", error=str(e))
-        from data.news_client import _get_demo_news
-
-        return build_news_feed(_get_demo_news())
-
-
-def _build_persona_kpis(
-    persona_id: str,
-    region: str,
-    demand_df: pd.DataFrame | None = None,
-    weather_df: pd.DataFrame | None = None,
-) -> dbc.Row:
-    """Build persona-specific KPI cards from live demand/weather data."""
-    capacity = REGION_CAPACITY_MW.get(region, 50000)
-
-    # Extract real stats from demand data
-    peak_mw = None
-    avg_mw = None
-    min_mw = None
-    pct_of_capacity = None
-    if demand_df is not None and "demand_mw" in demand_df.columns:
-        # Drop NaN and non-positive values — EIA occasionally emits spurious
-        # zero rows (e.g. NYISO 2026-02-10 had 6 consecutive zero-hours)
-        # that collapse Demand Range to the full peak. Utility demand is
-        # never physically 0 or negative for any covered BA.
-        valid = demand_df.dropna(subset=["demand_mw"])
-        valid = valid[valid["demand_mw"] > 0]
-        if not valid.empty:
-            peak_mw = valid["demand_mw"].max()
-            avg_mw = valid["demand_mw"].mean()
-            min_mw = valid["demand_mw"].min()
-            pct_of_capacity = peak_mw / capacity * 100 if capacity > 0 else 0
-
-    # Fallback: read demand stats from Redis
-    if peak_mw is None:
-        actuals_redis = redis_get(f"wattcast:actuals:{region}")
-        if actuals_redis and actuals_redis.get("demand_mw"):
-            demand_vals = [v for v in actuals_redis["demand_mw"] if v is not None and v > 0]
-            if demand_vals:
-                peak_mw = max(demand_vals)
-                avg_mw = sum(demand_vals) / len(demand_vals)
-                min_mw = min(demand_vals)
-                pct_of_capacity = peak_mw / capacity * 100 if capacity > 0 else 0
-
-    # Extract weather stats. pandas ``.mean()`` is NaN-aware but returns NaN
-    # when every value is null — coerce that to ``None`` so the downstream
-    # ``is not None`` guards (and Redis fallback) behave consistently.
-    avg_wind = None
-    avg_solar = None
-    if weather_df is not None:
-        if "wind_speed_80m" in weather_df.columns:
-            mean = weather_df["wind_speed_80m"].mean()
-            avg_wind = float(mean) if pd.notna(mean) else None
-        if "shortwave_radiation" in weather_df.columns:
-            mean = weather_df["shortwave_radiation"].mean()
-            avg_solar = float(mean) if pd.notna(mean) else None
-
-    # Fallback: read weather stats from Redis. Gate each metric independently
-    # so a missing column in ``weather_df`` still triggers the Redis lookup
-    # for that metric (previously both had to be ``None`` for any fallback).
-    # Also filter None/NaN entries before averaging — Redis arrays can have
-    # gaps where Open-Meteo emitted nulls.
-    if avg_wind is None or avg_solar is None:
-        weather_redis = redis_get(f"wattcast:weather:{region}")
-        if weather_redis:
-            if avg_wind is None and "wind_speed_80m" in weather_redis:
-                vals = [v for v in weather_redis["wind_speed_80m"] if pd.notna(v)]
-                if vals:
-                    avg_wind = sum(vals) / len(vals)
-            if avg_solar is None and "shortwave_radiation" in weather_redis:
-                vals = [v for v in weather_redis["shortwave_radiation"] if pd.notna(v)]
-                if vals:
-                    avg_solar = sum(vals) / len(vals)
-
-    # Get backtest MAPE from cache if available
-    backtest_mape = None
-    backtest_rmse = None
-    for horizon in [168, 24, 720]:  # prefer 7-day, then 24h, then 30d
-        bt_key = (region, horizon, "xgboost", DEFAULT_BACKTEST_EXOG_MODE)
-        if bt_key in _BACKTEST_CACHE:
-            cached_result, _, _ = _BACKTEST_CACHE[bt_key]
-            if "metrics" in cached_result:
-                backtest_mape = cached_result["metrics"].get("mape")
-                backtest_rmse = cached_result["metrics"].get("rmse")
-                break
-
-    # Fallback: read from Redis if in-memory cache is empty
-    if backtest_mape is None:
-        for horizon in [168, 24, 720]:
-            bt_redis = redis_get(
-                f"wattcast:backtest:{DEFAULT_BACKTEST_EXOG_MODE}:{region}:{horizon}"
-            )
-            if bt_redis is None:
-                bt_redis = redis_get(f"wattcast:backtest:{region}:{horizon}")
-            if bt_redis and "metrics" in bt_redis:
-                xgb_metrics = bt_redis["metrics"].get("xgboost", {})
-                if xgb_metrics:
-                    backtest_mape = xgb_metrics.get("mape")
-                    backtest_rmse = xgb_metrics.get("rmse")
-                    break
-
-    # Compute derived metrics
-    reserve_margin_pct = (100.0 - pct_of_capacity) if pct_of_capacity is not None else None
-    demand_range = (peak_mw - min_mw) if peak_mw is not None and min_mw is not None else None
-
-    # Wind capacity factor (approximate: avg_wind / rated_wind)
-    wind_cf = None
-    if avg_wind is not None:
-        from config import WIND_CUTOUT_SPEED_MPH
-
-        wind_cf = min(avg_wind / WIND_CUTOUT_SPEED_MPH * 100, 100.0)
-
-    # Solar capacity factor (approximate: avg_irradiance / rated_irradiance)
-    solar_cf = None
-    if avg_solar is not None:
-        from config import SOLAR_RATED_IRRADIANCE
-
-        solar_cf = avg_solar / SOLAR_RATED_IRRADIANCE * 100
-
-    # Estimate price from utilization (merit-order approximation)
-    from config import PRICING_BASE_USD_MWH
-
-    price_estimate = None
-    if pct_of_capacity is not None:
-        utilization = pct_of_capacity / 100
-        if utilization < 0.70:
-            price_estimate = PRICING_BASE_USD_MWH
-        elif utilization < 0.90:
-            price_estimate = PRICING_BASE_USD_MWH * (1 + (utilization - 0.70) * 5)
-        else:
-            price_estimate = PRICING_BASE_USD_MWH * (2 + (utilization - 0.90) * 20)
-
-    # Format values
-    peak_str = f"{int(peak_mw):,} MW" if peak_mw is not None else "No data"
-    avg_str = f"{int(avg_mw):,} MW" if avg_mw is not None else "No data"
-    cap_str = f"{pct_of_capacity:.0f}% of capacity" if pct_of_capacity is not None else ""
-    mape_str = f"{backtest_mape:.1f}%" if backtest_mape is not None else "No data"
-    mape_dir = "positive" if backtest_mape is not None and backtest_mape < 5 else "negative"
-    rmse_str = f"{int(backtest_rmse):,} MW" if backtest_rmse is not None else "No data"
-
-    persona_kpis = {
-        "grid_ops": [
-            {
-                "label": "Peak Demand",
-                "value": peak_str,
-                "delta": cap_str,
-                "direction": "negative" if pct_of_capacity and pct_of_capacity > 80 else "neutral",
-            },
-            {
-                "label": "Reserve Margin",
-                "value": f"{reserve_margin_pct:.0f}%"
-                if reserve_margin_pct is not None
-                else "No data",
-                "delta": "Below 15% is tight",
-                "direction": "negative"
-                if reserve_margin_pct is not None and reserve_margin_pct < 15
-                else "positive"
-                if reserve_margin_pct is not None
-                else "neutral",
-            },
-            {
-                "label": "Forecast Error",
-                "value": mape_str,
-                "delta": f"Walk-forward MAPE ({DEFAULT_BACKTEST_EXOG_MODE})",
-                "direction": mape_dir,
-            },
-            {
-                "label": "Demand Range",
-                "value": f"{int(demand_range):,} MW" if demand_range is not None else "No data",
-                "delta": "Peak - Min",
-                "direction": "neutral",
-            },
-        ],
-        "renewables": [
-            {
-                "label": "Wind CF",
-                "value": f"{wind_cf:.0f}%" if wind_cf is not None else "No data",
-                "delta": "Capacity factor",
-                "direction": "positive" if wind_cf is not None and wind_cf > 25 else "neutral",
-            },
-            {
-                "label": "Solar CF",
-                "value": f"{solar_cf:.0f}%" if solar_cf is not None else "No data",
-                "delta": "Capacity factor",
-                "direction": "positive" if solar_cf is not None and solar_cf > 15 else "neutral",
-            },
-            {
-                "label": "Avg Wind",
-                "value": f"{avg_wind:.1f} mph" if avg_wind is not None else "No data",
-                "delta": "80m hub height",
-                "direction": "neutral",
-            },
-            {
-                "label": "Avg Solar",
-                "value": f"{avg_solar:.0f} W/m\u00b2" if avg_solar is not None else "No data",
-                "delta": "Shortwave radiation",
-                "direction": "neutral",
-            },
-        ],
-        "trader": [
-            {
-                "label": "Est. Price",
-                "value": f"${price_estimate:.0f}/MWh" if price_estimate is not None else "No data",
-                "delta": "Merit-order estimate",
-                "direction": "negative"
-                if price_estimate is not None and price_estimate > 100
-                else "neutral",
-            },
-            {
-                "label": "Peak Demand",
-                "value": peak_str,
-                "delta": cap_str,
-                "direction": "neutral",
-            },
-            {
-                "label": "Avg Demand",
-                "value": avg_str,
-                "delta": f"Range: {int(demand_range):,} MW" if demand_range is not None else "",
-                "direction": "neutral",
-            },
-            {
-                "label": "Forecast Error",
-                "value": mape_str,
-                "delta": f"Walk-forward MAPE ({DEFAULT_BACKTEST_EXOG_MODE})",
-                "direction": mape_dir,
-            },
-        ],
-        "data_scientist": [
-            {
-                "label": "XGBoost MAPE",
-                "value": mape_str,
-                "delta": f"Target: <5% ({DEFAULT_BACKTEST_EXOG_MODE})",
-                "direction": mape_dir,
-            },
-            {
-                "label": "RMSE",
-                "value": rmse_str,
-                "delta": "Walk-forward backtest",
-                "direction": "neutral",
-            },
-            {
-                "label": "Peak Demand",
-                "value": peak_str,
-                "delta": cap_str,
-                "direction": "neutral",
-            },
-            {
-                "label": "Demand Range",
-                "value": f"{int(demand_range):,} MW" if demand_range is not None else "No data",
-                "delta": "Max variability",
-                "direction": "neutral",
-            },
-        ],
-    }
-    kpis = persona_kpis.get(persona_id, persona_kpis["grid_ops"])
-    return build_kpi_row(kpis)
 
 
 def _predict_single_fold(
